+++ RetriX.Shared.Test/project.json
@@ -10,7 +10,6 @@
     "PCLCrypto": "2.0.147",
     "PCLStorage": "1.0.2",
     "System.IO.Compression.ZipFile": "4.3.0",
-    "System.IO.FileSystem.Primitives": "4.3.0",
     "xunit": "2.2.0",
     "xunit.runner.visualstudio": "2.2.0"
   }
