+++ RetriX.Shared/StreamProviders/IStreamProvider.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 
 namespace RetriX.Shared.FileProviders
 {
-    public interface IFileProvider : IDisposable
+    public interface IStreamProvider : IDisposable
     {
         Task<IEnumerable<string>> ListEntriesAsync();
         Task<Stream> GetFileStreamAsync(string path, FileAccess accessType);
